make invalid UTF-8 round-trip through String
Modifies JS_NewStringLen's conversion from UTF-8 to UTF-16 to generate
unpaired low surrogates upon encountering non-decodable bytes, and
JS_ToCStringLen2's reverse conversion to decode that particular subset
of unpaired low surrogates into the corresponding bytes.

This is the so-called "UTF-8B" approach, and is precedented by e.g.
Python's PEP 383 <https://www.python.org/dev/peps/pep-0383/>.

This ensures that arbitrary byte data can be recovered even after being
stored in a JS String, which is important as many interfaces which may
represent arbitrary bytes (scriptArgs, std.getenv(), os.readdir(), &c.)
expose this data as a string, mangling values that do not decode as
valid UTF-8 in the process.

Unfortunately, it is as a result *not feasible* to maintain the ability
to create strings containing those unpaired surrogates by encoding them
(normatively illegally) in UTF-8 format--so-called "WTF-8"--as that
would entail having two byte streams encoding the same 16-bit data; it
would not be possible to recover the distinction between them.

My judgement is that the impact of this limitation is fairly low as it
affects only interactions with the C API, and it is comparatively less
common to manipulate potentially-malformed UTF-16, mostly involving
interfaces to legacy systems that assume Unicode is a 2-byte encoding.
Moreover, doing so is *already* cumbersome; although strings can be
extracted as CESU-8, the existing C interfaces are generally all UTF-8
based and do not present strings as sequences of 16-bit code units.