Fix js_debugger_eval can't eval non first lexical variable issue
For lexical variable, it is group and stored by the scope level, you can
only see the lexical variables that is little or equal your scope index
(inside your scope). But debugger is different, it must know all lexical
variables in the context.
So we define and use DEBUG_SCOP_INDEX to add all lexical variables for
debugger to eval it.

Signed-off-by: pengyaozong <pengyaozong@xiaomi.com>